+++ src/builders/fromNodePackageJson.node.ts
@@ -47,7 +47,7 @@ export class ToolBuilder {
   }
 
   // Current implementation does not return `undefined` yet, but it is an option for future implementation.
-  // To prevent breaking changes, it is declared to return `undefined`.
+  // To prevent future breaking changes, it is declared to return `undefined`.
   makeTool (data: PackageJson): Models.Tool | undefined {
     const [name, vendor] = typeof data.name === 'string'
       ? splitNameGroup(data.name)